Move BindToObject to a separate file
